daemon: Move `data_set_t` and related functions to its own file.
This is part of the legacy support for collectd 5 metrics. Because "data
sets" are a global resource, this has to remain within the daemon.
